// frontend/src/app/services/gateway-service-api/model/tokenResponse.ts
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { UserRolDto } from './userRolDto';


/**
 * 
 */
export interface TokenResponse { 
    access_token?: string;
    expires_in?: number;
    token_type?: string;
    roles?: Array<UserRolDto>;
    refresh_token?: string;
    expires_in_refresh?: number;
}